[SQL] Extract MultipleAggregationFn from BeamAggregationTransforms
